updated dotmailer detector; renamed to dotdigital
